V13 to main

+++ src/discordBot/commands/admin/delete_course.js
@@ -1,23 +1,29 @@
-const { updateGuide, findCategoryName, removeCourseFromDb, findCourseNickNameFromDbWithCourseCode } = require("../../services/service");
+const { updateGuide, findCategoryWithCourseName, removeCourseFromDb, findCourseNickNameFromDbWithCourseCode } = require("../../services/service");
 const { courseAdminRole } = require("../../../../config.json");
+const { confirmChoiceNoInteraction } = require("../../services/message");
 
 const execute = async (message, args, models) => {
   if (message.member.permissions.has("ADMINISTRATOR")) {
     let courseName = args.join(" ");
     const guild = message.guild;
 
-    let courseString = findCategoryName(courseName, guild);
-    let category = guild.channels.cache.find(c => c.type === "GUILD_CATEGORY" && c.name.toLowerCase() === courseString.toLowerCase());
+    const confirm = await confirmChoiceNoInteraction(message, "Delete course: " + courseName, guild);
+    if (!confirm) {
+      return;
+    }
+
+    let category = findCategoryWithCourseName(courseName, guild);
     if (!category) {
       const courseNickName = await findCourseNickNameFromDbWithCourseCode(courseName, models.Course);
       if (courseNickName) {
         courseName = String(courseNickName.dataValues.name);
-        courseString = findCategoryName(courseName, guild);
-        category = guild.channels.cache.find(c => c.type === "GUILD_CATEGORY" && c.name.toLowerCase() === courseString.toLowerCase());
+        category = findCategoryWithCourseName(courseName, guild);
       }
     }
 
+
     if (!category) return message.reply(`Error: Invalid course name: ${courseName}.`);
+
     await Promise.all(guild.channels.cache
       .filter(c => c.parent === category)
       .map(async channel => await channel.delete()),
